+++ src/plone/jsonapi/routes/providers/folders.py
@@ -3,7 +3,7 @@
 from plone.jsonapi.routes import add_plone_route
 
 # CRUD
-from plone.jsonapi.routes.api import get_items
+from plone.jsonapi.routes.api import get_batched
 from plone.jsonapi.routes.api import create_items
 from plone.jsonapi.routes.api import update_items
 from plone.jsonapi.routes.api import delete_items
@@ -17,12 +17,7 @@ from plone.jsonapi.routes.api import url_for
 def get(context, request, uid=None):
     """ get folders
     """
-    items = get_items("Folder", request, uid=uid, endpoint="folders")
-    return {
-        "url": url_for("folders"),
-        "count": len(items),
-        "items": items,
-    }
+    return get_batched("Folder", request, uid=uid, endpoint="folders")
 
 
 # CREATE